pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polygon-health-ai-query-engine"
version = "0.1.0"
description = "Polygon Health AI Query Engine"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.135.0",
  "uvicorn[standard]",
  "sqlalchemy[asyncio]",
  "asyncpg",
  "pydantic>=2",
  "pydantic-settings",
  "google-genai>=1.66.0",
  "openpyxl",
  "httpx",
  "langgraph>=0.2",
  "langgraph-checkpoint>=2",
  "langchain-core>=0.3",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "ruff",
  "mypy",
  "testcontainers[postgres]",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

